requirements.txt
discord.py
httpx
openai
pyyaml
anthropic
tenacity
